move core code to asapdiscovery
