+++ asapdiscovery/__init__.py
@@ -1,6 +1,4 @@
-"""Scripts and models for ML with COVID Moonshot data."""
-
-# Add imports here
+from . import data, docking, ml
 
 # Handle versioneer
 from ._version import get_versions
