--- covid_moonshot_ml/data/__init__.py
@@ -1,1 +1,0 @@
-# from . import dataset
