--- covid_moonshot_ml/datasets/__init__.py
@@ -1,1 +1,0 @@
-from . import fragalysis, moonshot, utils
