--- covid_moonshot_ml/nn/__init__.py
@@ -1,2 +1,0 @@
-from .models import *
-from .loss import *
