+++ scripts/dash-all-results-exploration.py
@@ -15,7 +15,7 @@ import json, argparse, os, sys
 repo_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(repo_path)
 
-from covid_moonshot_ml.docking.analysis import load_dataframes
+from covid_moonshot_ml.docking import load_dataframes
 
 
 parser = argparse.ArgumentParser(description="")
