+++ scripts/find_all_mcs.py
@@ -8,14 +8,13 @@ import sys
 
 sys.path.append(f"{os.path.dirname(os.path.abspath(__file__))}/../")
 from covid_moonshot_ml.datasets.utils import load_exp_from_sdf
-from covid_moonshot_ml.docking.docking import parse_xtal
-from covid_moonshot_ml.docking.mcs import (
+from covid_moonshot_ml.docking import parse_xtal
+from covid_moonshot_ml.docking import (
     rank_structures_openeye,
     rank_structures_rdkit,
 )
-from covid_moonshot_ml.schema import (
+from asapdiscovery.data.schema import (
     ExperimentalCompoundDataUpdate,
-    CrystalCompoundData,
     EnantiomerPairList,
 )
