+++ scripts/plot_docking_results.py
@@ -6,7 +6,7 @@ import plotly.express as px
 sys.path.append(
     f"{os.path.dirname(os.path.dirname(os.path.abspath(__file__)))}"
 )
-from covid_moonshot_ml.docking.analysis import DockingResults
+from covid_moonshot_ml.docking import DockingResults
 
 
 def get_args():
