+++ scripts/prep_proteins.py
@@ -22,8 +22,8 @@ from tempfile import NamedTemporaryFile
 import yaml
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from covid_moonshot_ml.schema import CrystalCompoundData
-from covid_moonshot_ml.modeling import (
+from asapdiscovery.data.schema import CrystalCompoundData
+from asapdiscovery.docking.modeling import (
     align_receptor,
     prep_receptor,
     du_to_complex,
@@ -32,12 +32,11 @@ from covid_moonshot_ml.modeling import (
 from covid_moonshot_ml.datasets import pdb
 from covid_moonshot_ml.datasets.utils import (
     save_openeye_pdb,
-    split_openeye_mol,
     edit_pdb_file,
     seqres_to_res_list,
     load_openeye_pdb,
 )
-from covid_moonshot_ml.docking.docking import parse_xtal
+from covid_moonshot_ml.docking import parse_xtal
 
 
 def check_completed(d):
