+++ scripts/run_docking_oe.py
@@ -14,7 +14,7 @@ import sys
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from covid_moonshot_ml.datasets.utils import load_openeye_sdf, save_openeye_sdf
-from covid_moonshot_ml.docking.docking import run_docking_oe
+from covid_moonshot_ml.docking import run_docking_oe
 
 
 def check_results(d):
@@ -263,7 +263,7 @@ def main():
 
     if args.exp_file:
         import json
-        from covid_moonshot_ml.schema import ExperimentalCompoundDataUpdate
+        from asapdiscovery.data.schema import ExperimentalCompoundDataUpdate
 
         ## Load compounds
         exp_compounds = [
